refactor(type): move the types file
