+++ src/Form/props.ts
@@ -1,7 +1,10 @@
 import type { PropType } from 'vue'
 import type { FormRulesMap } from 'element-plus/lib/components/form/src/form.type'
-import type { IComponentSize } from '../types/index'
-import type { IFormColumns, IFormMenuColumns } from './type'
+import type {
+  IComponentSize,
+  IFormColumns,
+  IFormMenuColumns,
+} from '../types/index'
 
 export default {
   modelValue: {
