+++ src/Table/index.ts
@@ -1,4 +1,3 @@
-export * from './type'
 import { objectDeepMerge } from '../utils/index'
 import ProTable from './Table'
 import props from './props'
