+++ src/composables/select.ts
@@ -9,8 +9,13 @@ import {
   watch,
 } from 'vue'
 import { isArray } from '../utils/index'
-import type { MaybeArray, UnknownObject } from '../types/index'
-import type { SelectConfig, SelectDataItem, SelectData } from '../Select/index'
+import type {
+  MaybeArray,
+  UnknownObject,
+  SelectConfig,
+  SelectDataItem,
+  SelectData,
+} from '../types/index'
 import type { ITreeSelectProps } from '../TreeSelect/index'
 import type TreeStore from 'element-plus/es/components/tree/src/model/tree-store'
 import type { FilterNodeMethodFunction } from 'element-plus/es/components/tree/src/tree.type'
