+++ src/types/index.ts
@@ -1,2 +1,8 @@
+export * from './crud'
+export * from './descriptions'
+export * from './form'
 export * from './public'
 export * from './router'
+export * from './select'
+export * from './table'
+export * from './tabs'
