+++ src/types/tabs.ts
